refactor user-avatar component
