+++ src/entities/profile/index.ts
@@ -1,2 +1,3 @@
 export type { Profile } from '../profile/model/types/profile'
+export * from './ui'
 export { ProfileCard } from './ui/profile-card/profile-card'
