--- src/shared/ui/user-avatar/user-avatar.tsx
@@ -1,15 +1,0 @@
-import { Avatar, AvatarFallback, AvatarImage, ReturnComponent } from '@/shared'
-
-type Props = {
-  className?: string
-  src?: string
-  userName?: string
-}
-export const UserAvatar = ({ className, src, userName }: Props): ReturnComponent => {
-  return (
-    <Avatar className={className}>
-      <AvatarImage alt={'user-avatar'} src={src} />
-      <AvatarFallback className={'bg-Light-900'}>{userName?.[0] || 'U'}</AvatarFallback>
-    </Avatar>
-  )
-}
